+++ custom_components/nexo/const.py
@@ -1,2 +1,3 @@
 """Constants for the nexo integration."""
 DOMAIN = "nexo"
+CONF_EXT_COMMANDS = "ext_commands"