+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.6.3] - 03.09.2025
+
+### Added
+- Blinds / Covers Groups integration
+- Ext commands (Automation / Smart)
+
+### Fixed
+- Switches Groups integration
+
+
+
+## [1.6.0] - 31.08.2025
+
+### Added
+- Dimmers integration - On/Off / Brightness
+- LED integration - On/Off / Brightness
+- Dimmers, LED Groups integration - On/Off / Brightness
+- Switches Groups integration  - On/Off
+- Weather station integration - Temperature, Wind, Weather pictogram, Wind direction, Light direction, Forecast from Open-Meteo
+
+
 ## [1.2.0] - 19.03.2024
 
 ### Added